Little settings view refactor
